Rewrite and libtoolize
Not much of the original remains - organize this like a real project
with autotools and multiple source files (!). The core of the library
opens a file and splits it into blocks (1024/4096 byte sectors) and
chunks. The command line utilities all operate on the chunks. Plenty
of room for optimization and improvement but this should get the party
properly started, five years later.