Improved support for non-Latin-1 characters
It looks like FileMaker uses a code-page switching technique for
non-Latin-1 characters, with a control character (0x03 and 0x05 observed
in the wild) followed by a data character. The data characters that I
observed in test files happened to match the last byte of the
corresponding Unicode code points, so I added some glue logic to
translate them to UTF-8. It's possible this was a coincidence and this
logic is flawed, so more test files are needed to know for certain.